requirements.txt
cachetools==2.1.0
certifi==2018.8.24
chardet==3.0.4
falcon==1.4.1
google-api-core==1.4.1
google-auth==1.5.1
google-cloud==0.34.0
google-cloud-core==0.28.1
google-cloud-datastore==1.7.0
googleapis-common-protos==1.5.3
grpcio==1.15.0
hug==2.4.1
idna==2.7
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
python-mimeparse==1.6.0
pytz==2018.5
requests==2.19.1
rsa==4.0
six==1.11.0
urllib3==1.23
